INDY-1304: Additional reorganization of tests
